Adds no control code bug test
Includes a test to reproduce the bug #61. It shows that if a control
package is partially send (only header) without the control code then the
circuit can read an old flit that represents a valid control code. This is
likely to happen as the control code flit can be delayed due to
retransmissions.

Commit: Add

TaskNumber: #61